Tests/LibWeb/Text/expected/HTML/img-in-picture-fires-single-load-event.txt

img load #1
sentinel load
